add analyser

+++ mte/YCSB/utils/rate_limit.h
@@ -34,7 +34,8 @@ class RateLimiter {
     // refill tokens
     auto now = Clock::now();
     auto diff = std::chrono::duration_cast<Duration>(now - last_);
-    tokens_ = std::min(b_, tokens_ + diff.count() * r_ / 1000000000);
+    //tokens_ = std::min(b_, tokens_ + diff.count() * r_ / 1000000000);
+    tokens_ = std::min(static_cast<int64_t>(b_), static_cast<int64_t>(tokens_ + diff.count() * r_ / 1000000000));
     last_ = now;
 
     // check tokens
@@ -54,7 +55,8 @@ class RateLimiter {
     // refill tokens
     auto now = Clock::now();
     auto diff = std::chrono::duration_cast<Duration>(now - last_);
-    tokens_ = std::min(b_, tokens_ + diff.count() * r_ * TOKEN_PRECISION / 1000000000);
+    //tokens_ = std::min(b_, tokens_ + diff.count() * r_ * TOKEN_PRECISION / 1000000000);
+    tokens_ = std::min(static_cast<int64_t>(b_), static_cast<int64_t>(tokens_ + diff.count() * r_ * TOKEN_PRECISION / 1000000000));
     last_ = now;
 
     // set rate
